Implement train and model - Change from accuracy to error as a measurement during learning

+++ src/dataset.py
@@ -41,4 +41,6 @@ class EncodingToTensor:
             inplace=True,
         )
 
-        return torch.tensor(x_copy), torch.tensor(sample[1])
+        return torch.tensor(x_copy, dtype=torch.float32), torch.tensor(
+            sample[1], dtype=torch.float32
+        )
